DHBox-modules/requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
backports.ssl-match-hostname==3.4.0.2
docker-py==0.6.0
itsdangerous==0.24
manage.py==0.2.6
requests==2.4.3
six==1.8.0
websocket-client==0.21.0
wsgiref==0.1.2
